+++ cloud-annotations/src/test/resources/TestCommandContainer.java
@@ -1,0 +1,6 @@
+import cloud.commandframework.annotations.processing.CommandContainer;
+
+@CommandContainer
+public class TestCommandContainer {
+
+}