+++ cloud-annotations/src/test/resources/TestCommandContainer2.java
@@ -1,0 +1,10 @@
+import cloud.commandframework.annotations.AnnotationParser;
+import cloud.commandframework.annotations.processing.CommandContainer;
+
+@CommandContainer
+public class TestCommandContainer2 {
+
+    public TestCommandContainer2(final AnnotationParser parser) {
+        // ...
+    }
+}